ai/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git